Updated artifactStores snippet for cross-region
Under section 'Add a Cross-Region Action to a Pipeline (CLI)' Step 3, the id of the encryption key is 'ID' for both regions which appears to be the same KMS key which is not the case. For different AWS region, you need to create different KMS keys for the respective regions.