Remove line indicating no x-account CodeBuild
I was able to create a pipeline that invokes a CodeBuild project in another account.  I was able to do this via an action role that has a trust relationship with the pipeline account.  The line being removed in this pr indicates that this would not be possible and thus could dissuade users from trying this setup.